+++ driver/src/test/java/com/arangodb/BaseJunit5.java
@@ -20,11 +20,11 @@ class BaseJunit5 {
     protected static final DbName TEST_DB = DbName.of("java_driver_test_db");
     protected static final ArangoConfigProperties config = ConfigUtils.loadConfig();
     private static final List<ArangoDB> adbs = Arrays.asList(
-            new ArangoDB.Builder().loadProperties(config).useProtocol(Protocol.VST).build(),
-            new ArangoDB.Builder().loadProperties(config).useProtocol(Protocol.HTTP_VPACK).build(),
-            new ArangoDB.Builder().loadProperties(config).useProtocol(Protocol.HTTP_JSON).build(),
-            new ArangoDB.Builder().loadProperties(config).useProtocol(Protocol.HTTP2_VPACK).build(),
-            new ArangoDB.Builder().loadProperties(config).useProtocol(Protocol.HTTP2_JSON).build()
+            new ArangoDB.Builder().loadProperties(config).protocol(Protocol.VST).build(),
+            new ArangoDB.Builder().loadProperties(config).protocol(Protocol.HTTP_VPACK).build(),
+            new ArangoDB.Builder().loadProperties(config).protocol(Protocol.HTTP_JSON).build(),
+            new ArangoDB.Builder().loadProperties(config).protocol(Protocol.HTTP2_VPACK).build(),
+            new ArangoDB.Builder().loadProperties(config).protocol(Protocol.HTTP2_JSON).build()
     );
 
     protected static Stream<ArangoDatabase> dbsStream() {